Reversed Polarity of Shutdown latch
The BMS MP_en which drives the shutdown latch is active low (ON) when there are NO faults present. When a critical DTC is present, or when a certain CANBUS message is recieved from the pedal box, the output will turn OFF (floats high), latching the relay open. If the output returns to normal (ON, active low), the relay will energize and close the shutdown loop